+++ texture_streamer_test.hpp
@@ -2,4 +2,4 @@
 
 #include <glm/glm.hpp>
 
-void texture_streamer_test_main(glm::vec2 const& sizes);
+void texture_streamer_test_main(glm::uvec2 const& sizes);
